chore(UsersModule): reorganize file structure
