+++ src/app/tracklists/state/tracklist-reducer.ts
@@ -2,7 +2,7 @@ import { Action, ActionReducer } from '@ngrx/store';
 import { List } from 'immutable';
 import { TRACKS_PER_PAGE } from 'src/constants';
 import { SearchActions } from 'src/app/search/search-actions';
-import { UserActions } from 'src/users/user-actions';
+import { UserActions } from 'src/app/users/user-actions';
 import { TrackData } from '../models/track';
 import { Tracklist, TracklistRecord } from '../models/tracklist';
 import { TracklistActions } from '../tracklist-actions';
