+++ src/app/users/components/user-card.scss
