+++ src/app/users/index.ts
@@ -1,0 +1,5 @@
+export { User, UserData, UserRecord } from './models';
+export { usersReducer, UsersState } from './state';
+export { UserActions } from './user-actions';
+export { UserService } from './user-service';
+export { UsersModule } from './users-module';
