+++ src/app/users/models/index.ts
@@ -1,0 +1,1 @@
+export { User, UserData, UserRecord } from './user';
