+++ src/app/users/state/index.ts
@@ -1,0 +1,1 @@
+export { usersReducer, UsersState } from './users-reducer';
