+++ src/app/users/user-effects.ts
@@ -12,7 +12,7 @@ import { Observable } from 'rxjs/Observable';
 import { AppState } from 'src/app';
 import { ApiService } from 'src/app/core';
 import { getCurrentTracklist, TracklistActions } from 'src/app/tracklists';
-import { getCurrentUser } from './reducers/selectors';
+import { getCurrentUser } from './state/selectors';
 import { UserActions } from './user-actions';
