+++ src/app/users/user-service.ts
@@ -5,7 +5,7 @@ import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs/Observable';
 import { AppState } from 'src/app';
 import { User } from './models/user';
-import { getCurrentUser } from './reducers/selectors';
+import { getCurrentUser } from './state/selectors';
 import { UserActions } from './user-actions';
